Amy/64 create challenge (#108)
* add floating action button to challenges fragment

* set up sliding screens for creating challenge

* set up three different fragments for each create challenge screen

* layout create challenge screens

* allow user to navigate through screens with cancel, back, and next buttons

* users can create challenges

* fix bugs - get user name and associated charity name and validate end date

* extract string resources and use preset values

* extract string resources and use preset values in story fragment

+++ app/src/main/java/com/example/una/models/Challenge.java
@@ -18,9 +18,8 @@ public class Challenge {
     Date startDate;
     Date endDate;
     double amountRaised;
-    long amountTarget;
+    double amountTarget;
     String associatedCharityEin;
-    // TODO remove after getting charity name from API
     String associatedCharityName;
     String frequency;
     String ownerId;
@@ -65,7 +64,7 @@ public class Challenge {
         }
 
         if (challengeFields.containsKey("target")) {
-            amountTarget = (long) challengeFields.get("target");
+            amountTarget = (double) challengeFields.get("target");
         }
 
         if (challengeFields.containsKey("associated_charity")) {
@@ -127,7 +126,7 @@ public class Challenge {
         return amountRaised;
     }
 
-    public long getChallengeAmountTarget() {
+    public double getChallengeAmountTarget() {
         return amountTarget;
     }
